Evaluator for release 0.5. It's working for negative numbers and decimal type of numbers. Yet to check in main.

+++ src/mritunjd/evaluator/lib/EvaluatorLib.java
@@ -10,40 +10,40 @@ public class EvaluatorLib {
         this.expression = expression;
     }
 
-    private int evaluateExpression(String expression) {
+    private double evaluateExpression(String expression) {
         String[] expressionParts = expression.trim().split(" ");
         List<String> operators = new ArrayList<String>();
-        List<Integer> operands = new ArrayList<Integer>();
+        List<Double> operands = new ArrayList<Double>();
         filterOperandsAndOperator(expressionParts, operators, operands);
         return calculateAll(operators, operands);
     }
 
-    private int calculateAll(List<String> operators, List<Integer> operands) {
+    private double calculateAll(List<String> operators, List<Double> operands) {
         Operations operationsMap = new Operations();
-        int operand1 = operands.get(0);
+        double operand1 = operands.get(0);
         for (int i = 0; i < operators.size(); i++) {
             String operator = operators.get(i);
-            int operand2 = operands.get(i + 1);
+            double operand2 = operands.get(i + 1);
             operand1 = operationsMap.performOperation(operator, operand1, operand2);
         }
         return operand1;
     }
 
-    public int evaluateExpression() {
+    public double evaluateExpression() {
         return evaluateExpression(this.expression);
     }
 
-    private void filterOperandsAndOperator(String[] expressionParts, List<String> operators, List<Integer> operands) {
+    private void filterOperandsAndOperator(String[] expressionParts, List<String> operators, List<Double> operands) {
         for (String part : expressionParts) {
             try {
-                operands.add(Integer.parseInt(part));
+                operands.add(Double.parseDouble(part));
             } catch (Exception ex) {
                 operators.add(part);
             }
         }
     }
 
-    private int startEvaluation(String inputExpression) {
+    private double startEvaluation(String inputExpression) {
         StringBuilder expression = new StringBuilder(inputExpression);
         if (expression.indexOf("(") > -1) {
             solveBrackets(expression);
@@ -52,7 +52,7 @@ public class EvaluatorLib {
         return evaluateExpression(expression.toString());
     }
 
-    public int startEvaluation() {
+    public double startEvaluation() {
         return startEvaluation(this.expression);
     }
 
@@ -68,7 +68,7 @@ public class EvaluatorLib {
             }
         }
         String expressionWithinBracket = expression.substring(openingBracketIndex + 1, closingBracketIndex);
-        int result = evaluateExpression(expressionWithinBracket);
-        expression.replace(openingBracketIndex, closingBracketIndex + 1, Integer.toString(result));
+        double result = evaluateExpression(expressionWithinBracket);
+        expression.replace(openingBracketIndex, closingBracketIndex + 1, Double.toString(result));
     }
 }
